refactor validator factory
